+++ Assets/Scripts/Interfaces/IDamagable.cs
@@ -2,6 +2,6 @@ using UnityEngine;
 
 public interface IDamagable
 {
-    public void Damage(int damage);
+    public void Damage(float damage);
 
 }
